Code Coverage Removal + TSX Explorations
Stil trying to see if I can move over to the built-in Node test runner instead of Jest, it didn't work the first time I tried moving to it. It doesn't have code coverage support, at least that I'm aware of, so I'm removing it in the Jest one first to help prevent accidental difference concerns.

I started moving everything to TypeScript, but I really don't like how complex Jest is to use with either TS or plain ESM, so I want to use `tsx` with the Node test runner instead, that's a very nice and simple setup, I really like it.

https://github.com/jestjs/jest/issues/13143