JSDom Module Instantiation
I was just about to revert back to Jest because this doesn't quite seem to be working properly either. The test seems to be getting stuck in a loop with calling the `.checkStatus()` method, when it wasn't doing that previously with the Jest test runner. I think it has something to do with my JSDom setup. Going to keep trying to fix it.