Gamepads Static Record
Moved this from the global scope as well, and into a static property instead. I really do like this about classes though, in the event where you need to track both something at the instance level, and at the global level. Scoping it to the class is better than polluting the global scope. I guess you could still do the same thing with a module, just by exporting getters and setters for that internal state that's managed. If you're already using a class for instancing though, then I feel like it makes sense to localize it to that as well though. Gotta say, classes are nice for state management, and in conjunction with instancing. If you don't need either of those though, then I feel like it isn't necessary, and maybe you can just use a function instead, or just a module. If it's a plain A -> B, use a function. If it's API is interactive, use a class. If it only has a single instance, why not just use object literals or a function to return the state structure? I guess that's what this project was doing previously, I think having instancing could be very valuable for this too though, like a multiplayer game, with split screen for example. They have a single set of data to manage all of the accessible controllers (static record), then the instances manage which ones they attach their game state to.